move to a mixin class
